+++ conreality_icons/CHANGELOG.md
@@ -6,6 +6,40 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.0.1 - 2019-07-06
+## 1.0.0 - 2019-07-08
 ### Added
 - `ConrealityIcons.version` getter
+- `ConrealityIcons.battery` icon
+- `ConrealityIcons.camera` icon
+- `ConrealityIcons.chat` icon
+- `ConrealityIcons.compass` icon
+- `ConrealityIcons.countdown` icon
+- `ConrealityIcons.dosimeter` icon
+- `ConrealityIcons.edgeDetection` icon
+- `ConrealityIcons.faceDetection` icon
+- `ConrealityIcons.faceRecognition` icon
+- `ConrealityIcons.filter` icon
+- `ConrealityIcons.font` icon
+- `ConrealityIcons.gamepad` icon
+- `ConrealityIcons.gps` icon
+- `ConrealityIcons.headset` icon
+- `ConrealityIcons.humidity` icon
+- `ConrealityIcons.icon` icon
+- `ConrealityIcons.joystick` icon
+- `ConrealityIcons.login` icon
+- `ConrealityIcons.logo` icon
+- `ConrealityIcons.map` icon
+- `ConrealityIcons.microphone` icon
+- `ConrealityIcons.motion` icon
+- `ConrealityIcons.nfc` icon
+- `ConrealityIcons.objectRecognition` icon
+- `ConrealityIcons.objectTracking` icon
+- `ConrealityIcons.plugin` icon
+- `ConrealityIcons.pulse` icon
+- `ConrealityIcons.replay` icon
+- `ConrealityIcons.script` icon
+- `ConrealityIcons.sensor` icon
+- `ConrealityIcons.temperature` icon
+- `ConrealityIcons.tracker` icon
+- `ConrealityIcons.video` icon
+- `ConrealityIcons.widget` icon